Shape intersections are norminetted

+++ include/utils.h
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/19 10:14:56 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/06 16:13:42 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/07 14:03:37 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	free_split_array(char **arr);
 char	*ft_strtrim_free(char *s, char *set);
 float	max3(float n1, float n2, float n3);
 float	min3(float n1, float n2, float n3);
-int		min(int a, int b);
+float		min(float a, float b);
 
 void	transform_ray(t_ray *transformed_ray, const t_ray *ray,
 			const t_shape *shape);
