+++ src/utils/math_utils.c
@@ -6,13 +6,13 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/02 17:21:50 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/02 17:22:52 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/07 14:03:25 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-int	min(int a, int b)
+float	min(float a, float b)
 {
 	if (a < b)
 		return (a);
